// src/app/model/gipher-model.ts
export class GipherModel {
    gipherId: string = '';
	url: string = '';
	recommendedBy: string = '';
	favoritedBy: string = '';
	title: string = '';
	userId: string = '';
	images: string = '';}
	// 	{
	// 		original:{
	// 			height :'',
	// 			width:'',
	// 			size:'',
	// 			url:'',
	// 			mp4_size:'',
	// 			mp4:'',
	// 			webp_size:'',
	// 			webp:'',
	// 			frames:'',
	// 			hash:''
	// 		},
	// 		downsized:{
	// 			height :'',
	// 			width:'',
	// 			size:'',
	// 			url:''
	// 		}

	// 	};
	// }
			// 'original_still':{
			// 	height :'',
			// 	width:'',
			// 	size:'',
			// 	url:''
			// },
			// 'original_mp4':{
			// 	height :'',
			// 	width:'',
			// 	size:'',
			// 	url:''
			// },
			// 'preview_gif':{
			// 	height :'',
			// 	width:'',
			// 	size:'',
			// 	url:''
			// },
			// 'downsized_large':{
			// 	height :'',
			// 	width:'',
			// 	size:'',
			// 	url:''
			// },
			// 'downsized_medium':{
			// 	height :'',
			// 	width:'',
			// 	size:'',
			// 	url:''
			// },
			// 'downsized_small':{
			// 	height :'',
			// 	width:'',
			// 	size:'',
			// 	url:''
			// },
			// 'downsized_still':{
			// 	height :'',
			// 	width:'',
			// 	size:'',
			// 	url:''
			// },
			// 'fixed_height':{
			// 	height :'',
			// 	width:'',
			// 	size:'',
			// 	url:'',
			// 	mp4_size:'',
			// 	mp4:'',
			// 	webp_size:'',
			// 	webp:''
			// },
			// 'fixed_height_downsampled':{
			// 	height :'',
			// 	width:'',
			// 	size:'',
			// 	url:'',
			// 	mp4_size:'',
			// 	mp4:'',
			// 	webp_size:'',
			// 	webp:''
			// },
			// 'fixed_height_small':{
			// 	height :'',
			// 	width:'',
			// 	size:'',
			// 	url:'',
			// 	mp4_size:'',
			// 	mp4:'',
			// 	webp_size:'',
			// 	webp:''
			// },
			// 'fixed_height_small_still':{
			// 	height :'',
			// 	width:'',
			// 	size:'',
			// 	url:'',
			// },
			// 'fixed_height_still':{
			// 	height :'',
			// 	width:'',
			// 	size:'',
			// 	url:'',
			// },
			// 'fixed_width':{
			// 	height :'',
			// 	width:'',
			// 	size:'',
			// 	url:'',
			// 	mp4_size:'',
			// 	mp4:'',
			// 	webp_size:'',
			// 	webp:''
			// },
			// 'fixed_width_downsampled':{
			// 	height :'',
			// 	width:'',
			// 	size:'',
			// 	url:'',
			// 	mp4_size:'',
			// 	mp4:'',
			// 	webp_size:'',
			// 	webp:''
			// },
			// 'fixed_width_small':{
			// 	height :'',
			// 	width:'',
			// 	size:'',
			// 	url:'',
			// 	mp4_size:'',
			// 	mp4:'',
			// 	webp_size:'',
			// 	webp:''
			// },
			// 'fixed_width_small_still':{
			// 	height :'',
			// 	width:'',
			// 	size:'',
			// 	url:''
			// },
			// 'fixed_width_still':{
			// 	height :'',
			// 	width:'',
			// 	size:'',
			// 	url:''
			// },
			// 'looping':{
			// 	height :'',
			// 	width:'',
			// 	size:'',
			// 	url:''
			// },
			// 'preview_webp':{
			// 	height :'',
			// 	width:'',
			// 	size:'',
			// 	url:''
			// },
			// 'hd':{
			// 	height :'',
			// 	width:'',
			// 	mp4_size:'',
			// 	mp4:'',
			// },
			// '480w_still':{
			// 	height :'',
			// 	width:'',
			// 	size:'',
			// 	url:''
			// }